refactor: Update API key management to use database
- Migrate createKey to use database repository
- Update deleteKey to use database operations
- Refactor manageKey to work with database
- Update updateKey for database compatibility
- Remove JSON file dependencies
