+++ server/functions/keys/manageKey.ts
@@ -1,5 +1,5 @@
 import Key from "../../types/api/keys/key";
-import { apiKeyRepository } from "../../db/repositories/apiKeyRepository";
+import { apiKeyRepository } from "../../db/repositories/apiKeyRepositoryMock";
 
 const listKeys = {
     list: async (): Promise<Key[]> => {
